Move messager's action and reducers to 'messager' dir
